restrict menu's buttons to respective users and public board layout and error messages presentation

+++ fastuga-laravel/app/Http/Controllers/api/ProductController.php
@@ -10,7 +10,8 @@ use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\App;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\StoreUpdateProductRequest;
+use App\Http\Requests\StoreProductRequest;
+use App\Http\Requests\UpdateProductRequest;
 use App\Http\Resources\ProductResource;
 use App\Models\OrderItems;
 use App\Models\Product;
@@ -40,21 +41,30 @@ class ProductController extends Controller
         return ProductResource::collection($sortedCollection);
     }
 
-    public function store(StoreUpdateProductRequest $request)
+    public function store(StoreProductRequest $request)
     {
-        $newProduct = Product::create($request->validated());
+        $validatedData = $request->validated();
         
-        if($request->hasFile('photo_url')){
-            $path = Storage::putFile('public/products',  $request->file('photo_url'));
-            $name = basename($path);
-            $newProduct["photo_url"] = $name;
-            $capitalizedName = ucfirst($newProduct["name"]);
-            $newProduct["name"] = $capitalizedName;
-            $newProduct->save();
+        if($request->has('photo_url')){
+            $image_64 = $request["photo_url"];
+            $extension = explode('/', explode(':', substr($image_64, 0, strpos($image_64, ';')))[1])[1];   // .jpg .png, ... 
+            //  image data to decode (eg: data:image/png;base64,imgData..)
+            $replace = substr($image_64, 0, strpos($image_64, ',')+1); // 
+            // find substring to replace      
+            $image = str_replace($replace, '', $image_64); 
+            $image = str_replace(' ', '+', $image); 
+            $imageName = Str::random(16).'.'.$extension;
 
-            return new ProductResource($newProduct);
+            Storage::put('public/products/'.$imageName, base64_decode($image));
+            
+            $validatedData["photo_url"] = $imageName;
         }
 
+        $capitalizedName = ucfirst($validatedData["name"]);
+        $validatedData["name"] = $capitalizedName;
+
+        $newProduct = Product::create($validatedData);
+        return new ProductResource($newProduct);
     }
 
     public function show(Product $product)
@@ -62,13 +72,12 @@ class ProductController extends Controller
         return new ProductResource($product);
     }
 
-    public function update(StoreUpdateProductRequest $request, Product $product)
+    public function update(UpdateProductRequest $request, Product $product)
     {
         $validatedData = $request->validated();
 
         // Check if New Photo Uploaded
-        // if($request->hasFile('photo_url')){
-        if($request->has('photo_url')){
+        if($request->has('photo_url') && Str::length($request["photo_url"]) > 21){
             // Delete Existing Photo
             if(Storage::disk('public')->exists('products/'.$product->photo_url)) {
                 Storage::disk('public')->delete('products/'.$product->photo_url);
@@ -87,7 +96,6 @@ class ProductController extends Controller
             $imageName = Str::random(16).'.'.$extension;
             
             Storage::put('public/products/'.$imageName, base64_decode($image));
-
             $validatedData["photo_url"] = $imageName;
         }
